LayoutTests/http/tests/security/contentSecurityPolicy/1.1/plugintypes-invalid.html

<!DOCTYPE html>
<html>
<head>
<script src='../resources/multiple-iframe-plugin-test.js'></script>
<script>
var tests = [
    [false, 'plugin-types;'],
    [false, 'plugin-types       ;'],
    [false, 'plugin-types text'],
    [false, 'plugin-types text/'],
    [false, 'plugin-types /text'],
    [false, 'plugin-types text//plain'],
    [false, 'plugin-types text/plainapplication/nospace'],
    [true, 'plugin-types text application/x-webkit-test-netscape'],
];
</script>
</head>
<body onload="test()">
    <p>
        This tests our handling of invalid `plugin-types` CSP directives.
        Consider this test passing if each of the following frames contains
        either "PASS" or no text at all.
    </p>
